feat(Asset): Improve asset disk removal conditions

+++ server/src/api/socket/asset_manager/core.py
@@ -17,6 +17,7 @@ from aiohttp_security import authorized_userid
 from .... import auth
 from ....app import app, sio
 from ....db.models.asset import Asset
+from ....db.models.asset_rect import AssetRect
 from ....db.models.user import User
 from ....logs import logger
 from ....state.asset import asset_state
@@ -182,7 +183,9 @@ async def assetmgmt_rm(sid: str, data: int):
 def clean_filehash(file_hash: str):
     if (ASSETS_DIR / file_hash).exists():
         no_assets = Asset.get_or_none(file_hash=file_hash) is None
-        if no_assets:
+        no_shapes = AssetRect.get_or_none(src=f"/static/assets/{file_hash}") is None
+        print(no_assets, no_shapes)
+        if no_assets and no_shapes:
             logger.info(f"No asset maps to file {file_hash}, removing from server")
             (ASSETS_DIR / file_hash).unlink()
